Adjust leader not only when we adding brokers but also when removing
Actually it is done separately from adding/removing replicas as a separate step now.
So even if replication factor is ok-ish but leaders are completely out of balance,
the tool may generate some reassignments.
It now prints some stats (number of partition a broker leads included) to explain
why the changes were made.